backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-filter==24.3
Pillow==11.1.0
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
